+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/FrameworkExtension.php
@@ -543,6 +543,16 @@ class FrameworkExtension extends Extension
             if (is_dir($dir = $container->getParameter('kernel.root_dir').'/Resources/translations')) {
                 $dirs[] = $dir;
             }
+            if (class_exists('Symfony\Component\Validator\Validator')) {
+                $r = new \ReflectionClass('Symfony\Component\Validator\Validator');
+
+                $dirs[] = dirname($r->getFilename()).'/Resources/translations';
+            }
+            if (class_exists('Symfony\Component\Form\Form')) {
+                $r = new \ReflectionClass('Symfony\Component\Form\Form');
+
+                $dirs[] = dirname($r->getFilename()).'/Resources/translations';
+            }
 
             // Register translation resources
             if ($dirs) {